Animation for Empty View

+++ TodoList/Views/ListView.swift
@@ -11,24 +11,32 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture(perform: {
-                        withAnimation(.snappy) {
-                            listViewModel.updateItem(item: item)
-                        }
-                    })
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture(perform: {
+                                withAnimation(.snappy) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            })
+                    }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
+            
         }
         .navigationTitle("Todo List 📝")
         .toolbar(content: {
-            ToolbarItem(placement: ToolbarItemPlacement.topBarLeading, content: {
-                EditButton()
+            ToolbarItemGroup(placement: ToolbarItemPlacement.topBarLeading, content: {
+                if !listViewModel.items.isEmpty {
+                    EditButton()
+                }
             })
-            
             
             ToolbarItemGroup(placement: .topBarTrailing, content: {
                 NavigationLink {
@@ -40,7 +48,7 @@ struct ListView: View {
         })
     }
     
-  
+    
 }
 
 #Preview {
